reorganize scripts
